SmellFactory introduced instead of RefactoringFactory. Generic Analysed class with its own Builder replaces ANode and AGroup
