+++ client/src/app/refactoring/dialog-analysis/dialog-analysis.component.ts
@@ -2,8 +2,6 @@ import { Component, OnInit } from '@angular/core';
 import { DynamicDialogRef } from 'primeng/dynamicdialog';
 import { DynamicDialogConfig } from 'primeng/dynamicdialog';
 import { AnalyserService } from '../analyser.service';
-import { GraphService } from "../../graph/graph.service";
-import { MessageService } from 'primeng/api';
 import { Smell } from "../../graph/model/smell";
 import { Principle } from '../../graph/model/principles';
