+++ client/src/app/refactoring/smell-factory.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { SmellFactoryService } from './smell-factory.service';
+
+describe('SmellFactoryService', () => {
+  let service: SmellFactoryService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(SmellFactoryService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
