+++ client/src/app/refactoring/smell.ts
@@ -114,81 +114,3 @@ export class GroupSmellObject {
         return descr;
     }
 }
-
-
-export class NoApiGatewaySmellObject extends GroupSmellObject {
-
-    constructor(group:joint.shapes.microtosca.Group) {
-        super("NoAPiGatewaySmell", group);
-    }
-
-    getDescription(){
-        let msg = "The node "
-        this.getNodeBasedCauses().forEach(node=>{
-            msg+= ` ${node.getName()}`
-        })
-        msg += " is accessed by external users without an Api Gateway."
-        return msg;
-    }
-}
-
-export class EndpointBasedServiceInteractionSmellObject extends SmellObject {
-
-    constructor() {
-        super("EndpointBasedServiceInterationSmell");
-    }
-
-    getDescription(): string {
-        var descr = "";
-        this.getLinkBasedCauses().forEach(link => {
-            let source = <joint.shapes.microtosca.Root>link.getSourceElement();
-            let target = <joint.shapes.microtosca.Root>link.getTargetElement();
-
-            descr += `Interaction from ${source.getName()} to ${target.getName()}.\n`;
-        })
-        return descr;
-    }
-}
-
-export class SharedPersistencySmellObject extends SmellObject {
-    constructor() {
-        super("Shared persistency");
-    }
-}
-export class WobblyServiceInteractionSmellObject extends SmellObject {
-
-    constructor() {
-        super("Wobbly service interaction");
-    }
-}
-
-export class SingleLayerTeamsSmellObject extends GroupSmellObject {
-
-    constructor(group:joint.shapes.microtosca.SquadGroup) {
-        super("Single-layer teams", group);
-    }
-
-    getDescription(): string {
-        let services = new Set<joint.shapes.microtosca.Root>();
-        this.getLinkBasedCauses().forEach(link => { services.add(<joint.shapes.microtosca.Root>link.getSourceElement()) });
-        return `${this.group.getName()} may lack some skills since there are services that access data owned by other teams: ${Array.from(services).map((service) => service.getName()).join(", ")}.`;
-    }
-}
-
-export class MultipleServicesInOneContainerSmellObject extends SmellObject {
-    constructor() {
-        super("Multiple services in one container")
-    }
-}
-
-export class TightlyCoupledTeamsSmell extends GroupSmellObject {
-
-    constructor(group:joint.shapes.microtosca.SquadGroup) {
-        super("Tightly coupled teams", group);
-    }
-
-    getDescription(): string {
-        return `${this.getNodeBasedCauses().map((n) => n.getName()).join(", ")} node${this,this.getNodeBasedCauses().length != 1 ? "s are" : " is"} more coupled to other teams than the current owner.`;
-    }
-
-}
