+++ client/src/app/refactoring/smells/multiple-services-in-one-container.ts
@@ -1,0 +1,7 @@
+import { SmellObject } from "../smell";
+
+export class MultipleServicesInOneContainerSmellObject extends SmellObject {
+    constructor() {
+        super("Multiple services in one container")
+    }
+}
