+++ client/src/app/refactoring/smells/shared-persistency.ts
@@ -1,0 +1,7 @@
+import { SmellObject } from "../smell";
+
+export class SharedPersistencySmellObject extends SmellObject {
+    constructor() {
+        super("Shared persistency");
+    }
+}
