+++ client/src/app/refactoring/smells/wobbly-service-interaction.ts
@@ -1,0 +1,8 @@
+import { SmellObject } from "../smell";
+
+export class WobblyServiceInteractionSmellObject extends SmellObject {
+
+    constructor() {
+        super("Wobbly service interaction");
+    }
+}
